% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABias_fun.R
\name{ABias_fun}
\alias{ABias_fun}
\title{Compute the Squared Average Bias (Squared ABias)}
\usage{
ABias_fun(modelFit, truth)
}
\arguments{
\item{modelFit}{A matrix of fitted probabilities of dimensions (number of observations x
number of simulations).}

\item{truth}{The true probabilities of the data generating process at \eqn{x_i}, i.e.,
\eqn{\Lambda (x_i^T \beta)}.}
}
\value{
A numeric scalar value.
}
\description{
Computes the squared average bias (Squared ABias) of an estimator
in a Monte Carlo simulation study.
}
\keyword{PredictingBlackSwans}
