% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMSPE_fun.R
\name{AMSPE_fun}
\alias{AMSPE_fun}
\title{Compute the Average Mean Squared Predicion Error (AMSPE) of an estimator}
\usage{
AMSPE_fun(modelFit, truth)
}
\arguments{
\item{modelFit}{A matrix of fitted probabilities of dimensions (number of observations x
number of simulations).}

\item{truth}{The true probabilities of the data generating process at \eqn{x_i}, i.e.,
\eqn{\Lambda (x_i^T \beta)}.}
}
\value{
A numeric scalar value.
}
\description{
Computes the Average Mean Squared Prediction Error (AMSPE) of an estimator
in a Monte Carlo simulation study.
}
\keyword{PredictingBlackSwans}
