% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVar_fun.R
\name{AVar_fun}
\alias{AVar_fun}
\title{Compute the Average Variance (AVariance) of an estimator}
\usage{
AVar_fun(modelFit)
}
\arguments{
\item{modelFit}{A matrix of fitted probabilities of dimensions (number of observations x
number of simulations).}
}
\value{
A numeric scalar value.
}
\description{
Computes the Average Variance (AVariance) of an estimator
in a Monte Carlo simulations study.
}
\keyword{PredictingBlackSwans}
