% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_nodewise_totalerr.R
\name{cv_nodewise_totalerr}
\alias{cv_nodewise_totalerr}
\title{Compute the total cross-validated error for the nodewise regression}
\usage{
cv_nodewise_totalerr(c, dataselects, x, lambdas, K)
}
\arguments{
\item{c}{Column of the response in the nodewise regression.}

\item{dataselects}{Fold index for the cross-validation.}

\item{x}{Matrix of predictors.}

\item{lambdas}{Sequence of regularization parameters.}

\item{K}{Number of folds for the cross-validation.}
}
\value{
A (Number of lambdas x Number of folds) matrix of
cross-validated errors (error on the discarded fold).
}
\description{
Compute the total cross-validated error for the nodewise regression
}
