% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despLassoLowComp.R
\name{despLassoLowComp}
\alias{despLassoLowComp}
\title{Compute the Desparsified Lasso Estimator for a Low Dimensional Component}
\usage{
despLassoLowComp(x, y, cvfolds, lowSel = seq(1, ncol(x)), parallel = TRUE,
  ncores = getOption("mc.cores", 2L))
}
\arguments{
\item{x}{Matrix of predictors.}

\item{y}{Response variable.}

\item{cvfolds}{Number of folds for the cross-validation for the initial
estimator.}

\item{lowSel}{Low dimensional selection of variables of interest for which
the desparsified Lasso will yield estimates.}

\item{parallel}{Should parallel computing be used when possible?}

\item{ncores}{Number of cores to be used when parallel is TRUE.}
}
\description{
Compute the Desparsified Lasso Estimator for Logistic Regression for a low
dimensional subset of variables of interest. The square-root Lasso with
Belloni, Chernozhukov & Wang (2011) tuning parameter is used.
}
