% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despLassoPaper.R
\name{despLassoPaper}
\alias{despLassoPaper}
\title{Compute the Desparsified Lasso Estimator as in the original paper by van de Geer (2014)}
\usage{
despLassoPaper(x, y, nodewise = c("cv", "sqrtLasso"), lambda = c("BCW",
  "VdG"), cvfolds = 5, parallel = TRUE, ncores = getOption("mc.cores",
  2L))
}
\arguments{
\item{x}{Matrix of predictors.}

\item{y}{Response variable.}

\item{nodewise}{Method to use for the nodewise Lasso: either
'cv' for Lasso with cross-validation or 'sqrtLasso' for the
square root nodewise Lasso or c('cv', 'sqrtLasso') for both estimators
of the approximate inverse matrix.}

\item{lambda}{Tuning parameter rule for the square-root nodewise Lasso. Either
BCW for the rule after Bernoulli, Chernozhukov & Wang (2011) or VdG for Van de Geer (2014).}

\item{cvfolds}{Number of folds for the cross-validation for both the initial
estimator and for the CV-nodewise-Lasso (if this is chosen).}

\item{parallel}{Should parallel computing be used when possible?}

\item{ncores}{Number of cores to be used when parallel is TRUE.}
}
\description{
Compute the Desparsified Lasso Estimator for Logistic Regression with the
scaling parameter sigmahat as in the original paper (van de Geer (2014)) using
the "desparsified outer product of the gradient" of the loss function instead of
the Hessian used in our preferred implementation.
}
