% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFun.R
\name{diffFun}
\alias{diffFun}
\title{Compute time differences}
\usage{
diffFun(inputData, inputVars)
}
\arguments{
\item{inputData}{A data.table.}

\item{inputVars}{Variables for which the time differences will be calculated.}
}
\value{
Additional variables in the dataset. These have the names
of the input variables with a "_diff"-ending.
}
\description{
Compute time differences y_t - t_{t-1}.
}
\examples{
diffFun(data=dat, inputVars=c("gdp", "revenue"))
}
\keyword{PredictingBlackSwans}
