% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZresiduals.R
\name{getZresiduals}
\alias{getZresiduals}
\title{Compute the residuals of the nodewise Lasso regressions}
\usage{
getZresiduals(i, x, lambda)
}
\arguments{
\item{i}{Column index of the response.}

\item{x}{Matrix of predictors.}

\item{lambda}{Regularization parameter.}
}
\value{
Vector of residuals.
}
\description{
Compute the residuals of the nodewise Lasso regressions
}
