% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZresidualsSQRTL.R
\name{getZresidualsSQRTL}
\alias{getZresidualsSQRTL}
\title{Compute the residuals of the square-root Lasso regressions}
\usage{
getZresidualsSQRTL(j, x, lambda)
}
\arguments{
\item{x}{Matrix of predictors.}

\item{lambda}{Either 'BCW' for the simulations method
proposed by Belloni, Chernozhukov and Wang (2011) or 'VdG' for
the proposed method in van de Geer (2014).}

\item{i}{Column index for the response.}
}
\value{
Vector of residuals.
}
\description{
Compute the residuals of the square-root Lasso regressions
}
