% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthFun.R
\name{growthFun}
\alias{growthFun}
\title{Compute growth rates}
\usage{
growthFun(data, inputVars)
}
\arguments{
\item{data}{A data.table.}

\item{inputVars}{Variables for which the growth rates will be calculated.}
}
\value{
Additional variables in the data.table. These have the names
of the input variables with a "_gr"-ending.
}
\description{
Compute growth rates as (y_t - y_{t-1}) / y_{t-1}.
}
\examples{
growthFun(data=dat, inputVars=c("gdp", "revenue"))
}
\keyword{PredictingBlackSwans}
