% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamiltonFilter.R
\name{hamiltonFilter}
\alias{hamiltonFilter}
\title{Implementation of the Hamilton (2017) filter}
\usage{
hamiltonFilter(inputData, inputVar, h = 3)
}
\arguments{
\item{inputData}{A data.table.}

\item{inputVar}{The variable for detrending.}

\item{h}{Horizon for which we build a prediction.}
}
\value{
A new data.table with an additional variable. This variable has the name
of the input variable with a '_dt'-ending.
}
\description{
Compute the deviations from trend for a time series using the Hamilton filter for each
country separately.
}
\keyword{PredictingBlackSwans}
