% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceMeanSePlot.R
\name{inferenceMeanSePlot}
\alias{inferenceMeanSePlot}
\title{Plot the standard errors of the estimate against the estimates}
\usage{
inferenceMeanSePlot(path)
}
\arguments{
\item{path}{Path to look for the input-files.}
}
\description{
Plot the Monte-Carlo standard errors of the estimate (y-axis) against the
Monte-Carlo estimates (x-axis) to assess the drivers of the inference results
regarding the worse coverage with simultaneouly better power properties of the tests.
}
\keyword{PredictingBlackSwans}
