% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceSim.R
\name{inferenceSim}
\alias{inferenceSim}
\title{Inference Simulation Study}
\usage{
inferenceSim(path = getwd(), parallel = T, ncores = getOption("mc.cores",
  2L), seed = 912)
}
\arguments{
\item{path}{Path to export the results.}

\item{parallel}{Should parallel computing be used? Note: It only works for UNIX systems.}

\item{ncores}{How many cores should be used for parallel computing?}

\item{seed}{Seed for reproducibility purposes.}
}
\description{
Replicate the results of our simulations part regarding inference. IMPORTANT NOTE: This
function is thought to be run as a script. See the details.
}
\details{
This function may take to long to run for computers with few kernels or
for Windows-computers. Therefore we suggest to run this function as an script to split
the computation of the simulations in several days.
}
\examples{
inferenceSim()
}
\keyword{PredictingBlackSwans}
