% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceSimMainPaper.R
\name{inferenceSimMainPaper}
\alias{inferenceSimMainPaper}
\title{Main function for the inference simulation study
using the "desparsified outer product of the gradient"}
\usage{
inferenceSimMainPaper(path = getwd(), n = 100, p = 150, rho = 0.5,
  nSim = 100, nomSize = 0.05, cvfolds = 5, seed = 182, parallel = T,
  ncores = getOption("mc.cores", 2L))
}
\arguments{
\item{path}{Path to export the results.}

\item{n}{Number of observations.}

\item{p}{Number of predictors.}

\item{rho}{Correlation parameter of the Toeplitz covariance matrix.}

\item{nSim}{Number of simulations.}

\item{nomSize}{Nominal size (type I error a.k.a. alpha).}

\item{cvfolds}{Number of folds for the cross-validations.}

\item{seed}{Seed for replication purposes.}

\item{parallel}{Should parallel computing be used? Note: It only works for UNIX systems.}

\item{ncores}{How many cores should be used for parallel computing?}
}
\description{
Computes all the results for the simulation study using the "desparsified outer product
of the gradient" as in the original paper (van de Geer (2014)).
}
\keyword{PredictingBlackSwans}
