% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceSimPrintResults.R
\name{inferenceSimPrintResults}
\alias{inferenceSimPrintResults}
\title{Print the inference simulation results in Latex format}
\usage{
inferenceSimPrintResults(inPath, outPath)
}
\arguments{
\item{inPath}{Path to look for the input-files.}

\item{outPath}{Path to export the results.}
}
\description{
Print the results of our simulation study for inference in Latex format as in the
master thesis.
}
\keyword{PredictingBlackSwans}
