% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsFun.R
\name{interactionsFun}
\alias{interactionsFun}
\title{Compute interactions terms}
\usage{
interactionsFun(inputData, inputVars)
}
\arguments{
\item{inputData}{A data.table.}

\item{inputVars}{Variables for which the interactions will be computed.}
}
\value{
Additional variables in the dataset. These have the names
of the pairs input variables with an "_i_" in between.
}
\description{
Compute interactions terms
}
\examples{
interactionsFun(data=dat, inputVars=c("gdp", "revenue"))
}
\keyword{PredictingBlackSwans}
