% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagFun.R
\name{lagFun}
\alias{lagFun}
\title{Compute lags of variables}
\usage{
lagFun(inputData, inputVars, maxLag)
}
\arguments{
\item{inputData}{A data.table.}

\item{inputVars}{Variables for which the lags will be computed.}

\item{maxLag}{Number up to which lags will be computed. E.g. if maxLag = 3
then the first, second and third lag will be calculated.}
}
\value{
Additional variables in the data.table. These will have the names
of the input variables with an "_jL"-ending, with j = 1, ..., maxAve.
}
\description{
Compute lags of variables
}
\examples{
lagFun(inputData=dat, inputVars=c("gdp", "revenue"), maxLag=3)
}
\keyword{PredictingBlackSwans}
