% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCCV_Analysis.R
\name{MCCV_Analysis}
\alias{MCCV_Analysis}
\title{Monte Carlo Cross-Validation Analysis for the performance comparison}
\usage{
MCCV_Analysis(mccvnumber = 100, cvfolds = 5, seed = 813, parallel = T,
  ncores = 2L)
}
\arguments{
\item{mccvnumber}{Number of Monte Carlo Cross-Validation runs.}

\item{cvfolds}{Number of folds for the cross-validation with the Lasso.}

\item{seed}{Seed for reproducibility.}

\item{parallel}{Should paralle computing be used? (Note: Only for UNIX computers).}

\item{ncores}{Number of cores.}
}
\description{
Main function to reproduce the results of our paper regarding the
performance comparison by means of Monte Carlo Cross-Validation (MCCV).
}
\keyword{PredictingBlackSwans}
