% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{Auxiliary function to compute the standard deviation with factor (1 / n)}
\usage{
mysd(y)
}
\arguments{
\item{y}{A vector.}
}
\value{
The standard deviation of the vector y with factor (1 / n) instead
of the default in base R (1 / (n- 1)).
}
\description{
Auxiliary function to compute the standard deviation with factor (1 / n)
}
