% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodewise_cv.R
\name{nodewise_cv}
\alias{nodewise_cv}
\title{Compute the nodewise Lasso using K-fold cross-validation}
\usage{
nodewise_cv(x, parallel = TRUE, ncores = 2L, lambda = "lambda.min",
  cvfolds = 5)
}
\arguments{
\item{x}{Predictor matrix.}

\item{parallel}{Should parallel computing be used?}

\item{ncores}{Number of cores for parallel computing.}

\item{lambda}{Either "lambda.1se" or "lambda.min" defined as
in the glmnet-Package. See e.g. ?glmnet::cv.glmnet.}

\item{cvfolds}{Number of folds for the cross-validation.}
}
\value{
The matrix of residuals of the nodewise Lasso regressions, i.e.
\eqn{Z = (Z_1, ..., Z_p)} with \eqn{Z_j \in R^n}.
}
\description{
Compute the nodewise Lasso using K-fold cross-validation and output
the matrix of residuals Z stemming from the nodewise regressions.
}
