% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodewise_sqrtlasso.R
\name{nodewise_sqrtlasso}
\alias{nodewise_sqrtlasso}
\title{Compute the nodewise Lasso using the square-root Lasso}
\usage{
nodewise_sqrtlasso(x, parallel = TRUE, ncores = 2L, lambda = "BCW")
}
\arguments{
\item{x}{Predictor matrix.}

\item{parallel}{Should parallel computing be used?}

\item{ncores}{Number of cores for parallel computing.}

\item{lambda}{Either 'BCW' for the simulations method
proposed by Belloni, Chernozhukov and Wang (2011) or 'VdG' for
the proposed method in van de Geer (2014).}
}
\value{
The matrix of residuals of the nodewise square-root Lasso regressions, i.e.
\eqn{Z = (Z_1, ..., Z_p)} with \eqn{Z_j \in R^n}.
}
\description{
Compute the nodewise Lasso using the square-root Lasso to
calculate the matrix of residuals Z stemming from the nodewise
regressions.
}
