% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrisProb.R
\name{plotCrisProb}
\alias{plotCrisProb}
\title{Plot Crisis Probabilities for each Country}
\usage{
plotCrisProb(crisDT, fullDT, countryName)
}
\arguments{
\item{crisDT}{A data table with the countries, years
and predicted probability.}

\item{fullDT}{The full data set.}

\item{countryName}{Name of the country to plot. It can also be "All".}
}
\value{
A ggplot.
}
\description{
Plot crisis probabilities for each country in the sample.
}
\keyword{PredictingBlackSwans}
