% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinEffects.R
\name{plotLinEffects}
\alias{plotLinEffects}
\title{Plot Linear Effects driving Crisis Probabilities}
\usage{
plotLinEffects(driverTable, fullDT, countryName)
}
\arguments{
\item{driverTable}{A data.table with the countries, years
and linear effects.}

\item{fullDT}{The full data set.}

\item{countryName}{Name of the country to plot.}
}
\value{
A ggplot.
}
\description{
Plot of the linear effects driving the crisis probabilities.
}
\keyword{PredictingBlackSwans}
