% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionSim.R
\name{predictionSim}
\alias{predictionSim}
\title{Prediction Simulation Study}
\usage{
predictionSim(path = getwd(), n = 100, pList = c(80, 150, 500),
  coefConfig = c("big", "small"), rho = 0.9, nSim = 100, cvfolds = 5,
  nVarInf = 20, parallel = TRUE, ncores = 2L, tau = 0.5, seed = 182)
}
\arguments{
\item{path}{Path to export the results.}

\item{n}{Number of observations.}

\item{nSim}{Number of simulations.}

\item{cvfolds}{Number of folds for the cross-validation.}

\item{nVarInf}{Number of variables for the infeasible logistic regression
estimator which knows a low-dimensional subset including the variables
with non-zero coefficient.}

\item{parallel}{Should parallel computing be used when possible?}

\item{ncores}{Number of cores for parallel computing.}

\item{tau}{Expected proportion of 1's in the data.}

\item{seed}{Seed for replication purposes.}

\item{p}{Number of covariates.}
}
\description{
Replicate the results of our simulations part regarding prediction accuracy.
}
\keyword{PredictingBlackSwans}
