% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqrt_lasso}
\alias{sqrt_lasso}
\title{Compute the square-root Lasso}
\usage{
sqrt_lasso(y, X, lambda)
}
\arguments{
\item{y}{Response variable.}

\item{X}{Matrix of regressors.}

\item{lambda}{Regularization parameter.}
}
\description{
Compute the square-root Lasso solution and its residuals for the nodewise Lasso.
}
\details{
The software is adapted from the Matlab-software provided by Belloni, Chernozhukov
and Wang (2011) in https://faculty.fuqua.duke.edu/~abn5/belloni-software.html
}
