% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAnalysis.R
\name{testAnalysis}
\alias{testAnalysis}
\title{Test Analysis}
\usage{
testAnalysis(mccvnumber = 100, cvfolds = 5, seed = 813, parallel = T,
  ncores = 2L)
}
\arguments{
\item{mccvnumber}{Number of Monte Carlo Cross-Validation runs.}

\item{cvfolds}{Number of folds for the cross-validation with the Lasso.}

\item{seed}{Seed for reproducibility.}

\item{parallel}{Should parallel computing be used? (Note: Only for UNIX computers)}

\item{ncores}{Number of cores for parallel computing}
}
\description{
Main function to reproduce the results of our paper regarding the
test analysis in the prediction part.
}
\keyword{PredictingBlackSwans}
