% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzingBS.R
\name{analyzingBS}
\alias{analyzingBS}
\title{Analyzing Black Swans}
\usage{
analyzingBS(path = getwd(), cvfolds = 5, seed = 813, parallel = T,
  ncores = 2L)
}
\arguments{
\item{path}{Path to export the results}

\item{cvfolds}{Number of folds for the cross-validation in the desparsified Lasso
for the initial estimator (Lasso).}

\item{seed}{Seed for reproducibility.}

\item{parallel}{Should parallel computing be used? (Only for UNIX computers)}

\item{ncores}{Number of cores for parallel-computing.}
}
\description{
Main function to reproduce the results of our paper regarding the
inference analysis.
}
\details{
A new directory "Inference_Analysis" will be created
where the results will be placed in.
}
\keyword{PredictingBlackSwans}
