% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despLasso.R
\name{despLasso}
\alias{despLasso}
\title{Compute the Desparsified Lasso Estimator}
\usage{
despLasso(x, y, nodewise = c("cv", "sqrtLasso"), lambda = c("BCW", "VdG"),
  cvfolds = 5, parallel = TRUE, ncores = getOption("mc.cores", 2L))
}
\arguments{
\item{x}{Matrix of predictors.}

\item{y}{Response variable.}

\item{nodewise}{Either 'cv' for the nodewise regression using the Lasso with
cross-validation or 'sqrtLasso' for the square-root Lasso.}

\item{lambda}{Tuning parameter for the square-root Lasso in the nodewise regressions.
Either 'BCW' for the proposal by Belloni, Chernozhukov and Wang (2011) or 'VdG' for
the proposal by van de Geer (2014).}

\item{cvfolds}{Number of folds for the cross-validation for both the initial
estimator and for the CV-nodewise-Lasso (if this is chosen).}

\item{parallel}{Should parallel computing be used when possible?}

\item{ncores}{Number of cores to be used when parallel is TRUE.}
}
\description{
Compute the Desparsified Lasso Estimator for Logistic Regression
}
