% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictingBS.R
\name{predictingBS}
\alias{predictingBS}
\title{Predicting Black Swans}
\usage{
predictingBS(path = getwd(), mccvnumber = 100, cvfolds = 5, seed = 813,
  parallel = T, ncores = 2L)
}
\arguments{
\item{path}{Path to export the results}

\item{mccvnumber}{Number of Monte Carlo Cross-Validations.}

\item{cvfolds}{Number of folds for the Lasso.}

\item{seed}{Seed for reproducibility.}

\item{parallel}{Should parallel computing be used? (Only for UNIX computers)}

\item{ncores}{Number of cores for parallel computing.}
}
\description{
Main function to reproduce the results of our paper regarding the
prediction analysis.
}
\details{
A new directory "Prediction_Analysis" will be created
where the results will be placed in.
}
\keyword{PredictingBlackSwans}
