% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionSim.R
\name{predictionSim}
\alias{predictionSim}
\title{Prediction Simulation Study}
\usage{
predictionSim(path = getwd(), n = 100, pList = c(80, 150, 500),
  coefConfig = c("big", "small"), rho = 0.9, nSim = 100, cvfolds = 5,
  parallel = TRUE, ncores = 2L, seed = 182)
}
\arguments{
\item{path}{Path to export the results.}

\item{n}{Number of observations.}

\item{pList}{Vector of number of covariates for each scenario.}

\item{coefConfig}{Either 'big' or 'small' or c('big', 'small') for
the scenarios corresponding to big or small coefficients.}

\item{rho}{Correlation strength of the Toeplitz covariance matrix.}

\item{nSim}{Number of simulations.}

\item{cvfolds}{Number of folds for the cross-validation.}

\item{parallel}{Should parallel computing be used when possible?}

\item{ncores}{Number of cores for parallel computing.}

\item{seed}{Seed for replication purposes.}
}
\description{
Replicate the results of our simulations part regarding prediction accuracy.
}
\keyword{PredictingBlackSwans}
